# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CommentRequest:
    properties:
      content:
        type: string
      parent_id:
        type: integer
    required:
    - content
    type: object
...
  dto.CreateListRequest:
    properties:
      description:
        type: string
      game_ids:
        items:
          type: integer
        type: array
      is_public:
        type: boolean
      thumbnail_img:
        type: string
      title:
        type: string
    required:
    - title
    type: object
  dto.CreateReviewRequest:
    properties:
      content:
        type: string
      game_id:
        type: integer
      img:
        type: string
      is_spoiler:
        type: boolean
      recommend:
        enum:
        - recommend
        - mixed
        - not_recommend
        type: string
      title:
        type: string
    required:
    - content
    - game_id
    - recommend
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      field:
        type: string
    type: object
...
  dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  dto.PaginatedResponse-array_dto_GameTrendingResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.GameTrendingResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationDTO'
      status:
        type: string
    type: object
  dto.PaginatedResponse-array_dto_ListTrendingResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ListTrendingResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationDTO'
      status:
        type: string
    type: object
  dto.PaginatedResponse-array_dto_NotificationResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.NotificationResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationDTO'
      status:
        type: string
    type: object
...
  dto.RateGameRequest:
    properties:
      game_id:
        type: integer
      rating:
        maximum: 5
        minimum: 0.5
        type: number
    required:
    - game_id
    - rating
    type: object
...
  dto.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      reset_token:
        type: string
    required:
    - new_password
    - reset_token
    type: object
...
  dto.SuccessResponse-array_dto_CommentResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CommentResponse'
        type: array
    type: object
  dto.SuccessResponse-dto_AuthResponse:
    properties:
      data:
        $ref: '#/definitions/dto.AuthResponse'
    type: object
  dto.SuccessResponse-dto_CommentResponse:
    properties:
      data:
        $ref: '#/definitions/dto.CommentResponse'
    type: object
...
  dto.SuccessResponse-dto_GameDetailResponse:
    properties:
      data:
        $ref: '#/definitions/dto.GameDetailResponse'
    type: object
  dto.SuccessResponse-dto_LikeGameResponse:
    properties:
      data:
        $ref: '#/definitions/dto.LikeGameResponse'
    type: object
  dto.SuccessResponse-dto_ListDetailResponse:
    properties:
      data:
        $ref: '#/definitions/dto.ListDetailResponse'
    type: object
  dto.SuccessResponse-dto_RateGameResponse:
    properties:
      data:
        $ref: '#/definitions/dto.RateGameResponse'
    type: object
  dto.SuccessResponse-dto_ReviewTrendingResponse:
    properties:
      data:
        $ref: '#/definitions/dto.ReviewTrendingResponse'
    type: object
  dto.SuccessResponse-dto_UserProfileResponse:
    properties:
      data:
        $ref: '#/definitions/dto.UserProfileResponse'
    type: object
...
  dto.SuccessResponse-dto_VerifyCodeResponse:
    properties:
      data:
        $ref: '#/definitions/dto.VerifyCodeResponse'
    type: object
  dto.SuccessResponse-string:
    properties:
      data:
        type: string
    type: object
  dto.UpdateListRequest:
    properties:
      description:
        type: string
      game_ids:
        items:
          type: integer
        type: array
      is_public:
        type: boolean
      thumbnail_img:
        type: string
      title:
        type: string
    type: object
  dto.UpdateReviewRequest:
    properties:
      content:
        type: string
      img:
        type: string
      is_spoiler:
        type: boolean
      recommend:
        enum:
        - recommend
        - mixed
        - not_recommend
        type: string
      title:
        type: string
    type: object
...
  dto.VerifyCodeRequest:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
...
  handlers.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.logoutRequest:
    properties:
      refresh_token:
        type: string
    type: object
  handlers.refreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handlers.registerRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Hệ thống API cho Gaming Box.
  title: Gaming Box API
  version: "1.0"
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Gửi yêu cầu khôi phục mật khẩu. (Luôn trả về thành công để bảo
        mật)
      parameters:
      - description: Email đăng ký
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Quên mật khẩu
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Đăng nhập bằng email và mật khẩu để nhận token
      parameters:
      - description: Thông tin đăng nhập
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Đăng nhập
      tags:
      - Authentication
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Xóa refresh token để đăng xuất
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.logoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Đăng xuất
      tags:
      - Authentication
  /auth/me:
    get:
      description: Lấy thông tin của người dùng đang đăng nhập
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_UserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy thông tin cá nhân
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Lấy access token mới bằng refresh token
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Làm mới token
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Đăng ký tài khoản mới bằng email và mật khẩu
      parameters:
      - description: Thông tin đăng ký
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Đăng ký tài khoản
      tags:
      - Authentication
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Đặt lại mật khẩu bằng token được cấp từ bước xác thực
      parameters:
      - description: Thông tin đổi mật khẩu
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Đặt lại mật khẩu
      tags:
      - Authentication
...
  /auth/steam/callback:
    get:
      description: Xử lý callback trả về từ Steam OpenID sau khi đăng nhập thành công
      responses:
        "200":
          description: HTML Script trả về Frontend
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Steam Callback
      tags:
      - Authentication (Steam)
  /auth/verify-code:
    post:
      consumes:
      - application/json
      description: Xác thực mã 6 số gửi qua email để lấy token đổi mật khẩu
      parameters:
      - description: Mã xác thực
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_VerifyCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Xác thực mã khôi phục
      tags:
      - Authentication
  /games/{game_id}/like:
    post:
      description: Thích hoặc Bỏ thích một game (Cần đăng nhập)
      parameters:
      - description: ID của Game
        in: path
        name: game_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_LikeGameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Thích hoặc Bỏ thích Game
      tags:
      - Likes
  /games/{id}:
    get:
      description: Lấy thông tin đầy đủ của một game dựa trên ID
      parameters:
      - description: ID của Game
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_GameDetailResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Lấy chi tiết Game
      tags:
      - Games
  /games/popular:
    get:
      description: Lấy danh sách game có đánh giá cao nhất
      parameters:
      - default: 1
        description: Trang hiện tại
        in: query
        name: page
        type: integer
      - default: 12
        description: Số lượng mỗi trang
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_GameTrendingResponse'
      summary: Danh sách Game phổ biến
      tags:
      - Games
  /games/rate:
    post:
      consumes:
      - application/json
      description: Đánh giá sao cho một game (Cần đăng nhập)
      parameters:
      - description: Thông tin đánh giá
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RateGameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_RateGameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Đánh giá Game
      tags:
      - Games
  /games/search:
    get:
      description: Tìm kiếm game theo tên có phân trang
      parameters:
      - description: Từ khóa tìm kiếm
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: Trang hiện tại
        in: query
        name: page
        type: integer
      - default: 12
        description: Số lượng mỗi trang
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_GameTrendingResponse'
      summary: Tìm kiếm Game
      tags:
      - Games
  /games/trending:
    get:
      description: Lấy danh sách game phổ biến có phân trang
      parameters:
      - description: Trang hiện tại (Mặc định 1)
        in: query
        name: page
        type: integer
      - description: Số lượng mỗi trang (Mặc định 12)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_GameTrendingResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Lấy danh sách game thịnh hành
      tags:
      - Games
  /lists:
    post:
      consumes:
      - application/json
      description: Tạo danh sách game cá nhân (Cần đăng nhập)
      parameters:
      - description: Thông tin list
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateListRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_ListDetailResponse'
      security:
      - BearerAuth: []
      summary: Tạo List mới
      tags:
      - Lists
  /lists/{id}:
    delete:
      description: Xóa một danh sách game (Chỉ chủ sở hữu)
      parameters:
      - description: ID của List
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-string'
      security:
      - BearerAuth: []
      summary: Xóa List
      tags:
      - Lists
    get:
      description: Lấy thông tin và danh sách game trong một list
      parameters:
      - description: ID của List
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_ListDetailResponse'
      summary: Xem chi tiết List
      tags:
      - Lists
    put:
      consumes:
      - application/json
      description: Chỉnh sửa thông tin hoặc danh sách game trong list (Chỉ chủ sở
        hữu)
      parameters:
      - description: ID của List
        in: path
        name: id
        required: true
        type: integer
      - description: Thông tin cập nhật
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_ListDetailResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật List
      tags:
      - Lists
  /lists/{list_id}/like:
    post:
      description: Thích hoặc Bỏ thích một list (Cần đăng nhập)
      parameters:
      - description: ID của List
        in: path
        name: list_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_LikeGameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Thích hoặc Bỏ thích List
      tags:
      - Likes
  /lists/trending:
    get:
      description: Lấy danh sách list phổ biến có phân trang
      parameters:
      - description: Trang hiện tại (Mặc định 1)
        in: query
        name: page
        type: integer
      - description: Số lượng mỗi trang (Mặc định 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_ListTrendingResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Lấy danh sách List thịnh hành
      tags:
      - Lists
  /notifications:
    get:
      description: Lấy danh sách thông báo của người dùng hiện tại (Cần đăng nhập)
      parameters:
      - default: 1
        description: Trang hiện tại
        in: query
        name: page
        type: integer
      - default: 20
        description: Số lượng hiển thị trên mỗi trang
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_NotificationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách thông báo
      tags:
      - Notifications
  /notifications/{id}/read:
    post:
      description: Đánh dấu một thông báo cụ thể là đã đọc
      parameters:
      - description: ID của thông báo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Đánh dấu đã đọc một thông báo
      tags:
      - Notifications
  /notifications/read-all:
    post:
      description: Đánh dấu tất cả thông báo của người dùng hiện tại là đã đọc
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Đánh dấu đã đọc tất cả thông báo
      tags:
      - Notifications
  /reviews:
    post:
      consumes:
      - application/json
      description: Tạo một review cho game (Cần đăng nhập)
      parameters:
      - description: Thông tin review
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_ReviewTrendingResponse'
      security:
      - BearerAuth: []
      summary: Tạo Review mới
      tags:
      - Reviews
  /reviews/{id}:
    delete:
      description: Xóa một review (Chỉ chủ sở hữu)
      parameters:
      - description: ID của Review
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-string'
      security:
      - BearerAuth: []
      summary: Xóa Review
      tags:
      - Reviews
    put:
      consumes:
      - application/json
      description: Chỉnh sửa review đã viết (Chỉ chủ sở hữu mới có quyền)
      parameters:
      - description: ID của Review
        in: path
        name: id
        required: true
        type: integer
      - description: Thông tin cập nhật
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_ReviewTrendingResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật Review
      tags:
      - Reviews
  /reviews/{id}/comments:
    get:
      description: Lấy tất cả bình luận của một review
      parameters:
      - description: ID của Review
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-array_dto_CommentResponse'
      summary: Lấy danh sách bình luận
      tags:
      - Reviews
    post:
      consumes:
      - application/json
      description: Bình luận vào một review (Cần đăng nhập)
      parameters:
      - description: ID của Review
        in: path
        name: id
        required: true
        type: integer
      - description: Nội dung bình luận
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_CommentResponse'
      security:
      - BearerAuth: []
      summary: Thêm bình luận
      tags:
      - Reviews
  /reviews/{review_id}/like:
    post:
      description: Thích hoặc Bỏ thích một review (Cần đăng nhập)
      parameters:
      - description: ID của Review
        in: path
        name: review_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_LikeGameResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Thích hoặc Bỏ thích Review
      tags:
      - Likes
  /reviews/trending:
    get:
      description: Lấy danh sách review phổ biến có phân trang
      parameters:
      - description: Trang hiện tại (Mặc định 1)
        in: query
        name: page
        type: integer
      - description: Số lượng mỗi trang (Mặc định 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_ReviewTrendingResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Lấy danh sách Review thịnh hành
      tags:
      - Reviews
  /users/diary:
    get:
      description: Lấy danh sách nhật ký chơi game của người dùng
      parameters:
      - description: ID người dùng
        in: query
        name: id
        type: integer
      - default: 1
        description: Trang
        in: query
        name: page
        type: integer
      - default: 20
        description: Giới hạn
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_DiaryEntry'
      summary: Xem Nhật ký chơi game
      tags:
      - Users
  /users/follow:
    post:
      consumes:
      - application/json
      description: Thực hiện follow hoặc unfollow một người dùng khác (Cần đăng nhập)
      parameters:
      - description: Thông tin follow
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.FollowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_FollowResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Follow / Unfollow Người dùng khác
      tags:
      - Users
  /users/me:
    get:
      description: Lấy thông tin profile đầy đủ của người dùng (Cần đăng nhập)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_UserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy Profile Người dùng
      tags:
      - Users
  /users/me/followers:
    get:
      description: Lấy danh sách những người dùng đang follow current user
      parameters:
      - default: 1
        description: Trang hiện tại
        in: query
        name: page
        type: integer
      - default: 20
        description: Số lượng hiển thị trên mỗi trang
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_UserSummary'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách người theo dõi
      tags:
      - Users
  /users/me/following:
    get:
      description: Lấy danh sách những người dùng mà current user đang follow
      parameters:
      - default: 1
        description: Trang hiện tại
        in: query
        name: page
        type: integer
      - default: 20
        description: Số lượng hiển thị trên mỗi trang
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_UserSummary'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy danh sách đang theo dõi
      tags:
      - Users
  /users/profile:
    get:
      description: Lấy thông tin profile đầy đủ của một người dùng dựa theo ID
      parameters:
      - description: ID của Người dùng
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_UserProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Xem Profile người dùng khác
      tags:
      - Users
  /users/stats:
    get:
      description: Xem thống kê về game đã chơi, đánh giá trung bình...
      parameters:
      - description: ID người dùng (Nếu không có sẽ lấy me)
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_UserStatsResponse'
      summary: Lấy thống kê người dùng
      tags:
      - Users
  /users/watchlist:
    get:
      description: Lấy danh sách game muốn chơi của người dùng
      parameters:
      - description: ID người dùng
        in: query
        name: id
        type: integer
      - default: 1
        description: Trang
        in: query
        name: page
        type: integer
      - default: 20
        description: Giới hạn
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedResponse-array_dto_GameSummary'
      summary: Xem Watchlist (Backlog)
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"