# internal/routes/routes.go
package routes

import (
	"net/http"
	"vault/be/config"
	"vault/be/internal/handlers"
	"vault/be/internal/middleware"

	"github.com/gin-gonic/gin"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
)

func SetupRouter(authHandler *handlers.AuthHandler, steamHandler *handlers.SteamHandler, userHandler *handlers.UserHandler, gameHandler *handlers.GameHandler, reviewHandler *handlers.ReviewHandler, listHandler *handlers.ListHandler, likeHandler *handlers.LikeHandler, notifHandler *handlers.NotificationHandler) *gin.Engine {
	if config.App.Env == "production" {
		gin.SetMode(gin.ReleaseMode)
	}

	r := gin.Default()
	r.Use(corsMiddleware())

	r.GET("/health", func(c *gin.Context) {
		c.JSON(http.StatusOK, gin.H{"status": "ok", "env": config.App.Env})
	})

	// Swagger UI Route
	r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	v1 := r.Group("/api/v1")
	auth := v1.Group("/auth")
	{
		auth.GET("/steam", steamHandler.LoginHandle)
		auth.GET("/steam/callback", steamHandler.CallbackHandle)
		auth.POST("/register", authHandler.Register)
		auth.POST("/login", authHandler.Login)
		auth.POST("/refresh", authHandler.RefreshToken)
		auth.POST("/forgot-password", authHandler.ForgotPassword)
		auth.POST("/verify-code", authHandler.VerifyCode)
		auth.POST("/reset-password", authHandler.ResetPassword)

		protected := auth.Group("")
		protected.Use(middleware.Authenticate())
		{
			protected.GET("/me", authHandler.Me)
			protected.POST("/logout", authHandler.Logout)
		}
	}

	users := v1.Group("/users")
	{
		users.GET("/profile", userHandler.GetProfile)
		users.GET("/stats", userHandler.GetStats)
		users.GET("/diary", userHandler.GetDiary)
		users.GET("/watchlist", userHandler.GetWatchlist)

		usersProtected := users.Group("")
		usersProtected.Use(middleware.Authenticate())
		{
			usersProtected.GET("/me", userHandler.Me)
			usersProtected.GET("/me/following", userHandler.GetFollowing)
			usersProtected.GET("/me/followers", userHandler.GetFollowers)
			usersProtected.POST("/follow", userHandler.FollowUser)
		}
	}

	games := v1.Group("/games")
	{
		games.GET("/trending", gameHandler.TrendingGames)
		games.GET("/popular", gameHandler.PopularGames)
		games.GET("/search", gameHandler.SearchGames)
		games.GET("/:id", gameHandler.GetGameDetail)

		gamesProtected := games.Group("")
		gamesProtected.Use(middleware.Authenticate())
		{
			gamesProtected.POST("/rate", gameHandler.RateGame)
		}
	}

	reviews := v1.Group("/reviews")
	{
		reviews.GET("/trending", reviewHandler.TrendingReviews)
		reviews.GET("/:id", reviewHandler.GetReviewDetail)
		reviews.GET("/:id/comments", reviewHandler.GetComments)

		reviewsProtected := reviews.Group("")
		reviewsProtected.Use(middleware.Authenticate())
		{
			reviewsProtected.POST("", reviewHandler.CreateReview)
			reviewsProtected.PUT("/:id", reviewHandler.UpdateReview)
			reviewsProtected.DELETE("/:id", reviewHandler.DeleteReview)
			reviewsProtected.POST("/:id/comments", reviewHandler.AddComment)
		}
	}

	lists := v1.Group("/lists")
	{
		lists.GET("/trending", listHandler.TrendingLists)
		lists.GET("/:id", listHandler.GetListDetail)

		listsProtected := lists.Group("")
		listsProtected.Use(middleware.Authenticate())
		{
			listsProtected.POST("", listHandler.CreateList)
			listsProtected.PUT("/:id", listHandler.UpdateList)
			listsProtected.DELETE("/:id", listHandler.DeleteList)
		}
	}

	likesGroup := v1.Group("")
	likesGroup.Use(middleware.Authenticate())
	{
		likesGroup.POST("/games/:id/like", likeHandler.LikeGame)
		likesGroup.POST("/reviews/:id/like", likeHandler.LikeReview)
		likesGroup.POST("/lists/:id/like", likeHandler.LikeList)
	}

	notifications := v1.Group("/notifications")
	notifications.Use(middleware.Authenticate())
	{
		notifications.GET("", notifHandler.GetNotifications)
		notifications.POST("/:id/read", notifHandler.MarkAsRead)
		notifications.POST("/read-all", notifHandler.MarkAllAsRead)
	}

	return r
}

func corsMiddleware() gin.HandlerFunc {
	return func(c *gin.Context) {
		c.Header("Access-Control-Allow-Origin", config.App.FrontEndUrl)
		c.Header("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE, OPTIONS")
		c.Header("Access-Control-Allow-Headers", "Origin, Content-Type, Authorization")
		c.Header("Access-Control-Allow-Credentials", "true")

		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(http.StatusNoContent)
			return
		}
		c.Next()
	}
}